Move theme to own package
