+++ package.json
@@ -28,8 +28,8 @@
         "publish": "pnpm run build && pnpm run changeset publish"
     },
     "devDependencies": {
+        "@solid-devtools/theme": "file:packages/theme",
         "@changesets/cli": "^2.26.2",
-        "@nothing-but/utils": "~0.11.1",
         "@playwright/test": "^1.39.0",
         "@total-typescript/ts-reset": "^0.5.1",
         "@types/node": "^20.8.10",
