+++ packages/theme/package.json
@@ -1,0 +1,46 @@
+{
+    "name": "@solid-devtools/theme",
+    "version": "0.0.0",
+    "license": "MIT",
+    "author": "Damian Tarnawski <gthetarnav@gmail.com>",
+    "contributors": [],
+    "homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/theme#readme",
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/thetarnav/solid-devtools.git"
+    },
+    "bugs": {
+        "url": "https://github.com/thetarnav/solid-devtools/issues"
+    },
+    "private": false,
+    "publishConfig": {
+        "access": "public"
+    },
+    "sideEffects": false,
+    "files": [
+        "./theme.ts"
+    ],
+    "type": "module",
+    "module": "./theme.ts",
+    "types": "./theme.ts",
+    "exports": {
+        "types": "./theme.ts",
+        "import": "./theme.ts"
+    },
+    "scripts": {
+        "dev": "echo \"No dev script\"",
+        "build": "echo \"No build script\"",
+        "test:unit": "echo \"No test:unit script\"",
+        "test:types": "tsc --noEmit"
+    },
+    "dependencies": {
+        "@nothing-but/utils": "~0.11.1"
+    },
+    "packageManager": "pnpm@8.9.0",
+    "keywords": [
+        "solid",
+        "devtools",
+        "css",
+        "unocss"
+    ]
+}
